uclibc: remove linuxthreads.new, rename linuxthreads.old
Linuxthreads.new isn't really useful with the existence
of NPTL/TLS for well supported architectures. There is no
reason to use LT.new for ARM/MIPS or other architectures
supporting NPTL/TLS. It is not available for noMMU architectures
like Blackfin or FR-V. To simplify the live of the few uClibc-ng
developers, LT.new is removed and LT.old is renamed to LT.
LINUXTHREADS_OLD -> UCLIBC_HAS_LINUXTHREADS

+++ uclibc/lib/contrib/uclibc/libpthread/linuxthreads/sysdeps/pthread/not-cancel.h
@@ -23,10 +23,10 @@
 /* Uncancelable open.  */
 #if defined __NR_openat && !defined __NR_open
 #define open_not_cancel(name, flags, mode) \
-	INLINE_SYSCALL (openat, 4, (int) (AT_FDCWD), (const char *) (name), \
+	INLINE_SYSCALL (openat, 4, AT_FDCWD, (const char *) (name), \
 		(flags), (mode))
 #define open_not_cancel_2(name, flags) \
-	INLINE_SYSCALL (openat, 3, (int) (AT_FDCWD), (const char *) (name), \
+	INLINE_SYSCALL (openat, 3, AT_FDCWD, (const char *) (name), \
 		(flags))
 #else
 #define open_not_cancel(name, flags, mode) \
